requirements.txt
streamlit
google-generativeai>=0.7.0
python-dotenv
pyPDF2
pillow
openpyxl
pandas